+++ test/methods/test_lemma.py
@@ -2,6 +2,7 @@ from test.test_data.lemma_data import lemma_data
 
 import pytest
 
+from wiktionary_de_parser.models import Lemma
 from wiktionary_de_parser.parser.parse_lemma import ParseLemma
 
 
@@ -10,4 +11,4 @@ class TestLemmaParsing:
     def test_parsing_lemmas(self, test_input, expected):
         wikitext = test_input
         result = ParseLemma.parse("Untitled", wikitext)
-        assert result == expected
+        assert result == Lemma(**expected)